+++ test/TestCases/Case09/Case09.02.test.js
@@ -1,4 +1,4 @@
-import {readJSONFile, testApolloArtifacts} from '../../testLib';
+import { checkFolderContainsFiles, readJSONFile, testApolloArtifacts } from '../../testLib';
 import fs from "fs";
 import {parseNeptuneEndpoint} from "../../../src/util.js";
 
@@ -7,9 +7,17 @@ const testDbInfo = parseNeptuneEndpoint(testCase.host + ':' + testCase.port);
 
 const outputFolderPath = './test/TestCases/Case09/output';
 describe('Validate Apollo Server Subgraph output artifacts', () => {
-    afterAll(async () => {
+    afterAll(() => {
         fs.rmSync(outputFolderPath, {recursive: true});
     });
+
+    checkFolderContainsFiles(outputFolderPath, [
+        `${testDbInfo.graphName}.output.neptune.schema.json`,
+        `${testDbInfo.graphName}.output.resolver.graphql.js`,
+        `${testDbInfo.graphName}.output.resolver.schema.json`,
+        `${testDbInfo.graphName}.output.schema.graphql`,
+        `${testDbInfo.graphName}.output.source.schema.graphql`
+    ]);
 
     testApolloArtifacts(outputFolderPath, testDbInfo, true);
 });